+++ lockfile_generator/src/pip.rs
@@ -30,17 +30,17 @@ impl Generator for Pip {
             "--dry-run",
         ]);
 
-        // Check if we got a requirements file or a setup.py/pyproject.toml.
-        let is_requirements_file =
-            manifest_path.file_name().and_then(|f| f.to_str()).map_or(false, |file_name| {
-                file_name == "requirements.in"
-                    || (file_name.starts_with("requirements") && file_name.ends_with(".txt"))
-            });
+        // Check if we got a local project file (setup.py/pyproject.toml).
+        // All other files will be treated as requirements files.
+        let is_local_project = manifest_path
+            .file_name()
+            .and_then(|f| f.to_str())
+            .map_or(false, |file_name| file_name == "setup.py" || file_name == "pyproject.toml");
 
-        if is_requirements_file {
-            command.arg("-r").arg(manifest_path);
-        } else {
+        if is_local_project {
             command.arg(".");
+        } else {
+            command.arg("-r").arg(manifest_path);
         }
 
         command